Divided the algorithm API into four headers for different types, so that one can include algorithms for a specific type without including algorithms for all types.
